Got characters moving around. Animations still needed

+++ src/entities/components.rs
@@ -4,13 +4,13 @@ use bevy::ecs::component::Component;
 #[derive(Component)]
 pub struct CharacterEntity {
     pub name: String,
-    pub id: u32,
+    pub id: u64,
 }
 
 #[derive(Component)]
 pub struct NPCEntity {
     pub name: String,
-    pub id: u32,
+    pub id: u64,
 }
 
 #[derive(Component)]
